# api/src/config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PulseOps V1 API",
    "description": "Enterprise Modular Operations Platform — Backend API with hot-drop module architecture, JWT authentication, and role-based access control.",
    "version": "1.0.0",
    "contact": {
      "name": "PulseOps Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4001/api",
      "description": "Local Development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /auth/login"
      }
    },
    "schemas": {
...
      "ModuleInfo": {
        "type": "object",
        "properties": {
          "moduleId": { "type": "string" },
          "name": { "type": "string" },
          "version": { "type": "string" },
          "description": { "type": "string" },
          "isCore": { "type": "boolean" },
          "enabled": { "type": "boolean" },
          "order": { "type": "integer" }
        }
      }
    }
  },
  "paths": {
...
    "/modules": {
      "get": {
        "tags": ["Modules"],
        "summary": "List all installed modules",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of installed modules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ModuleInfo" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/modules/available": {
      "get": {
        "tags": ["Modules"],
        "summary": "Discover available hot-drop modules from dist-modules/ folder",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "List of available modules from hot-drop folder" }
        }
      }
    },
    "/modules/{moduleId}": {
      "get": {
        "tags": ["Modules"],
        "summary": "Get a single module by ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "moduleId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Module details" },
          "404": { "description": "Module not found" }
        }
      },
      "delete": {
        "tags": ["Modules"],
        "summary": "Remove an installed module",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "moduleId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Module removed" },
          "400": { "description": "Cannot remove core module" }
        }
      }
    },
    "/modules/{moduleId}/install": {
      "post": {
        "tags": ["Modules"],
        "summary": "Install a module from the available list",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "moduleId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Module installed" },
          "409": { "description": "Module already installed" }
        }
      }
    },
    "/modules/{moduleId}/enable": {
      "post": {
        "tags": ["Modules"],
        "summary": "Enable an installed module",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "moduleId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Module enabled" }
        }
      }
    },
    "/modules/{moduleId}/disable": {
      "post": {
        "tags": ["Modules"],
        "summary": "Disable an installed module",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "moduleId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Module disabled" },
          "400": { "description": "Cannot disable core module" }
        }
      }
    },
...
  }
}